+++ index.js
@@ -4,15 +4,14 @@ const createModuleConfig = require('./src/module');
 const createOutputConfig = require('./src/output');
 const createPluginsConfig = require('./src/plugins');
 const createResolveConfig = require('./src/resolve');
-const resolveAbsolutePath = require('./src/utils/resolve-absolute-path');
-
+const applyWizardConfigDefaults = require('./apply-wizard-config-defaults');
 const WEBPACK_CONFIG = Symbol('webpack-wizard-config');
 
 const webpackWizard = (config) => {
   if (config[WEBPACK_CONFIG]) {
     return config;
   }
-  const wizardConfig = applyConfigDefaults(config);
+  const wizardConfig = applyWizardConfigDefaults(config);
   const webpackConfig = {};
   webpackConfig.output = createOutputConfig(wizardConfig);
   webpackConfig.entry = createEntryConfig(wizardConfig);
@@ -25,32 +24,5 @@ const webpackWizard = (config) => {
   webpackConfig[WEBPACK_CONFIG] = true;
   return webpackConfig;
 };
-
-const applyConfigDefaults = (wizardConfig = {}) => ({
-  isDev: wizardConfig.isDev || process.env.NODE_ENV === 'development',
-  isProd: wizardConfig.isProd || process.env.NODE_ENV === 'production',
-  devHost: wizardConfig.devHost || 'localhost',
-  devPort: wizardConfig.devPort || 3000,
-  env: wizardConfig.env || {},
-  input: applyInputDefaults(wizardConfig.input || {}),
-  output: applyOutputDefaults(wizardConfig.output || {})
-});
-
-const applyInputDefaults = (input) => ({
-  favicon: input.favicon || resolveAbsolutePath('favicon.ico'),
-  html: input.html || resolveAbsolutePath('index.html'),
-  js: input.js || resolveAbsolutePath('src/index.js'),
-  jsDev: input.jsDev || resolveAbsolutePath('src/index-dev.js'),
-  modules: input.modules || [ resolveAbsolutePath('src') ],
-  styles: input.styles || resolveAbsolutePath('src/styles'),
-  stylesGlobals: input.stylesGlobals || 'globals.scss'
-});
-
-const applyOutputDefaults = (output) => ({
-  directory: output.directory || resolveAbsolutePath('dist'),
-  css: output.css || 'styles.css',
-  html: output.html || 'index.html',
-  js: output.js || 'bundle.js'
-});
 
 module.exports = webpackWizard;
